Moved all utilities to helpers for testing
